fix(*): added the types
