+++ app/hooks/index.ts
@@ -1,4 +1,4 @@
 export { useConfigs } from "./useConfigs";
 export { useInfiniteScroll } from "./useInfiniteScroll";
 export { usePaginatedList } from "./usePaginatedList";
-export type { UsePaginatedListResult } from "./usePaginatedList";
+export type { UsePaginatedListResult } from "@/app/lib/types/pagination";
